+++ CHANGES.md
@@ -2,11 +2,43 @@
 
 Here you can see an overview of changes between each release.
 
+## Version 4.1.1_04
+
+Released on June 5th, 2020.
+
+* Added RClone to interact with Jackrabbit via webdav protocol.
+* Added feature to cleanup unused metadata XML TR file.
+* Upgraded `oxShibbolethIdp`.
+
+## Version 4.1.1_03
+
+Released on April 29th, 2020.
+
+* Added Jackrabbit/JCR support.
+
+## Version 4.1.1_02
+
+Released on April 2nd, 2020.
+
+* Disabled `shibwatcher` script.
+
+## Version 4.1.1_01
+
+Released on March 24th, 2020.
+
+* Conformed to Gluu Server v4.1.1.
+
 ## Version 4.1.0_01
 
 Released on March 5th, 2020.
 
 * Conformed to Gluu Server v4.1.
+
+## Version 4.0.1_06
+
+Released on April 2nd, 2020.
+
+* Disabled `shibwatcher` script.
 
 ## Version 4.0.1_05